+++ test/kube2e/util.go
@@ -11,18 +11,7 @@ import (
 	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/options"
 	"github.com/solo-io/go-utils/testutils/helper"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
-
-	"github.com/solo-io/go-utils/kubeutils"
-	"k8s.io/client-go/kubernetes"
 )
-
-func MustKubeClient() kubernetes.Interface {
-	restConfig, err := kubeutils.GetConfig("", "")
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	kubeClient, err := kubernetes.NewForConfig(restConfig)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	return kubeClient
-}
 
 // Check that everything is OK by running `glooctl check`
 func GlooctlCheckEventuallyHealthy(testHelper *helper.SoloTestHelper, timeoutInterval string) {
